+++ src/init_support.h
@@ -26,7 +26,9 @@ namespace forge
     std::vector<std::string> definitions;
     std::vector<std::string> macos_frameworks;
     std::vector<std::string> macos_libraries;
+    std::vector<std::string> macos_brew_packages;
     std::vector<std::string> linux_libraries;
+    std::vector<std::string> linux_apt_packages;
     std::vector<std::string> windows_libraries;
     std::vector<std::filesystem::path> references;
     std::map<std::string, BuildProfile> profiles;